+++ CHANGELOG.md
@@ -6,6 +6,24 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
+## 2025.03.30
+
+### Added
+
+- Designed abstract class for GameLibrary component
+
+### Updated
+
+- Changed design to include the following secondary methods:
+
+#### Secondary Methods Added
+
+selectAny()
+updateTime(Game game, int playtime)
+hasTime(int playtime)
+String toString()
+boolean equals(GameLibrary gl2)
+
 ## 2025.02.03
 
 ### Added